Extract LayoutConstraintEvaluator from PushPredicateIntoTableScan
